fix: resolve pkg-file-count violations for operator defaults PR
Consolidate operator-default runtime helpers and relocate tests so
pkg/cli, pkg/config, and pkg/service stay within the 15-file limit.
